SCRUM-115 fix: update way to baseApi
