+++ .storybook/withReduxDecorator.tsx
@@ -2,7 +2,7 @@ import { Provider } from 'react-redux'
 import { configureStore } from '@reduxjs/toolkit'
 import { StoryFn } from '@storybook/react'
 import { appSlice } from '@/src/shared/model/slices/appSlice'
-import { baseApi } from '@/src/shared/model/api/baseApi'
+import { baseApi } from '@/src/shared/model/api/base/baseApi'
 
 // 2. Настраиваем store для Storybook
 const storybookStore = configureStore({
