+++ src/app/posts/page.tsx
@@ -1,3 +1,4 @@
+'use client'
 import { RequireAuth } from '@/src/features/requireAuth/RequireAuth'
 import { AddPost } from '@/src/widgets/addPost/AddPost'
