+++ src/shared/model/api/base/baseQueryWithReauth.ts
@@ -1,31 +1,12 @@
 import { AUTH_KEYS } from '@/src/shared/lib/constants/auth-keys'
 import { handleError } from '@/src/shared/lib/handleError'
-import {
-  BaseQueryFn,
-  FetchArgs,
-  FetchBaseQueryError,
-  createApi,
-  fetchBaseQuery,
-} from '@reduxjs/toolkit/query/react'
+import { isTokens } from '@/src/shared/lib/isTokens'
+import { baseQuery } from '@/src/shared/model/api/base/baseQuery'
+import { BaseQueryFn, FetchArgs, FetchBaseQueryError } from '@reduxjs/toolkit/query/react'
 import { Mutex } from 'async-mutex'
 
 // create a new mutex
 const mutex = new Mutex()
-
-const baseQuery = fetchBaseQuery({
-  baseUrl: process.env.NEXT_PUBLIC_API_BASE_URL,
-  credentials: 'include',
-  prepareHeaders: headers => {
-    const token = localStorage.getItem(AUTH_KEYS.ACCESS_TOKEN)
-
-    if (token) {
-      headers.set('Authorization', `Bearer ${token}`)
-    }
-
-    return headers
-  },
-})
-//
 
 export const baseQueryWithReauth: BaseQueryFn<
   FetchArgs | string,
@@ -59,14 +40,10 @@ export const baseQueryWithReauth: BaseQueryFn<
         )
 
         // )as any //что бы не ругалась на типизацию
-        // console.log(refreshResult)
-        if (
-          typeof refreshResult.data === 'object' &&
-          refreshResult.data !== null &&
-          'accessToken' in refreshResult.data &&
-          refreshResult.data?.accessToken &&
-          typeof refreshResult.data.accessToken === 'string'
-        ) {
+        console.log(refreshResult)
+        debugger
+        if (refreshResult.data && isTokens(refreshResult.data)) {
+          debugger
           localStorage.setItem(AUTH_KEYS.ACCESS_TOKEN, refreshResult.data.accessToken)
           // retry the initial query
           result = await baseQuery(args, api, extraOptions)
@@ -88,26 +65,3 @@ export const baseQueryWithReauth: BaseQueryFn<
 
   return result
 }
-
-export const baseApi = createApi({
-  baseQuery: baseQueryWithReauth, // Используем кастомный baseQuery
-  endpoints: () => ({}),
-  reducerPath: 'inctagramApi',
-  tagTypes: [
-    'ME',
-    'POSTS',
-    'POST',
-    'FOLLOWING',
-    'COMMENTS',
-    'PROFILE',
-    'SESSIONS',
-    'PAYMENTS',
-    'NOTIFICATIONS',
-    'FEED',
-    'POST_LIKES',
-    'COMMENT_LIKES',
-    'ANSWER_LIKES',
-    'ANSWERS',
-    'MESSAGES',
-  ],
-})
