+++ src/shared/model/api/postsApi.ts
@@ -18,7 +18,7 @@ import {
   UpdatePostModel,
 } from '@/src/entities/post/types'
 import { PREVIEW_LIKES_LIMIT } from '@/src/shared/lib/constants/post'
-import { baseApi } from '@/src/shared/model/api/baseApi'
+import { baseApi } from '@/src/shared/model/api/base/baseApi'
 import { setAppError, setAppSuccess } from '@/src/shared/model/slices/appSlice'
 import { setLastPostId } from '@/src/shared/model/slices/postsSlice'
