--- src/shared/model/api/types.ts
@@ -1,3 +1,0 @@
-export type ModalSuccessType = {
-  type: 'error' | 'success'
-}
